% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FactorAnalysis.R
\name{FactorAnalysis}
\alias{FactorAnalysis}
\title{Factor analysis with non-random missing data}
\usage{
FactorAnalysis(
  Y,
  Cov = NULL,
  K = NULL,
  Miss.Mech,
  ind.samples = NULL,
  method = c("MSNIMBLE", "sIPW", "IPW"),
  max.miss = 0.5,
  BH.min = NULL,
  include.updates = T,
  n.quad = 100
)
}
\arguments{
\item{Y}{a \code{p} x \code{n} data matrix of log2-transformed metabolite intensities, where \code{p} = #of metabolites and \code{n} = #of samples. Missing values should be left as \code{NA}.}

\item{Cov}{a \code{n} x \code{r} matrix of nuisance covariates (i.e. intercept, sex, etc.).}

\item{K}{The number of latent covariates (i.e. C is a \code{n} x \code{K} matrix). If unspecified, it is estimated using sva::num.sv applied to the metabolites with complete data.}

\item{Miss.Mech}{The missingness mechansim object returned by \code{EstimateMissing}.}

\item{ind.samples}{A logical or numeric vector of samples to be considered in the analysis. For example, if nuisance covariates were only measured in a subset of the samples, this would be the samples with nuisance covariates. Default is no missing samples.}

\item{method}{The method used to estimate L. Can be one of "MSNIMBLE" (the default and recommended; uses both observed and missing data to improve power), "sIPW" (stabilized inverse probability weighting; discards missing data), or "IPW" (inverse probability weighting; discards missing data).}

\item{max.miss}{Maximum fraction of missing data a metabolite is allowed to have to be used in the estimate for the latent factors C. Defaults to 0.5.}

\item{n.quad}{Number of Gaussian quadrature points. Default, and recommended value, is 100.}
}
\value{
A list \item{C}{Estimates for the latent factors. The columns are orthogonal, orthogonal to Cov, and have squared norm \code{NROW(Cov)-NCOL(Cov)}.} \item{L}{Estimates for the latent loadings. Columns are orthogonal, and are arranged so that their squared norms are decreasing.} \item{Sigma}{The metabolite-specific residuals variances.} \item{Var.L}{A list of K x K variances of the rows of L.} \item{stand.eigen.values}{Standardized eigenvalue estimates of 1/#metabolites*L(C'C)L', which have been standardized \code{mean(Sigma)}.} \item{Perc.Var.Explained}{The percent metabolite level variance explained by the \code{K} latent factors.} \item{method}{The method used to estimate L.}
}
\description{
Perform factor analysis in metabolite data with non-random missing observations. The goal is to estimate L and C in the model Y = LC' + E, where we identify L and C using "PCA identification", i.e. C'C is proportional to the identity and L'L is diagonal with decreasing elements.
}
